# openapi.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/user": {"get": {"summary": "Get All Users", "operationId": "get_all_users_user_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get All Users User Get", "type": "array", "items": {"$ref": "#/components/schemas/ShowUser"}}}}}}}, "post": {"summary": "Create User", "operationId": "create_user_user_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/user/{user_id}": {"get": {"summary": "Get User By Id", "operationId": "get_user_by_id_user__user_id__get", "parameters": [{"required": true, "schema": {"title": "User Id"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowUser"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"summary": "Update User By Id", "operationId": "update_user_by_id_user__user_id__put", "parameters": [{"required": true, "schema": {"title": "User Id"}, "name": "user_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateUser"}}}, "required": true}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete User", "operationId": "delete_user_user__user_id__delete", "parameters": [{"required": true, "schema": {"title": "User Id"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai": {"get": {"summary": "Get All Ai", "operationId": "get_all_ai_ai_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get All Ai Ai Get", "type": "array", "items": {"$ref": "#/components/schemas/ShowAI"}}}}}}}, "post": {"summary": "Create Ai", "operationId": "create_ai_ai_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateAI"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/privacy": {"get": {"summary": "Get All Public Ai", "operationId": "get_all_public_ai_privacy_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get All Public Ai Privacy Get", "type": "array", "items": {"$ref": "#/components/schemas/AI"}}}}}}}}, "/ai/{ai_id}": {"get": {"summary": "Get Ai By Id", "operationId": "get_ai_by_id_ai__ai_id__get", "parameters": [{"required": true, "schema": {"title": "Ai Id"}, "name": "ai_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ShowAI"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"summary": "Update Ai By Id", "operationId": "update_ai_by_id_ai__ai_id__put", "parameters": [{"required": true, "schema": {"title": "Ai Id"}, "name": "ai_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateAI"}}}, "required": true}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete Ai", "operationId": "delete_ai_ai__ai_id__delete", "parameters": [{"required": true, "schema": {"title": "Ai Id"}, "name": "ai_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/files/pythonscript": {"post": {"summary": "Create Script File", "operationId": "create_script_file_ai_files_pythonscript_post", "parameters": [{"required": true, "schema": {"title": "Model Id"}, "name": "model_id", "in": "query"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_create_script_file_ai_files_pythonscript_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/ai/files/modelfiles": {"post": {"summary": "Create Model File", "operationId": "create_model_file_ai_files_modelfiles_post", "parameters": [{"required": true, "schema": {"title": "Ai Id"}, "name": "ai_id", "in": "query"}], "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_create_model_file_ai_files_modelfiles_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Main", "operationId": "main__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AI": {"title": "AI", "required": ["ai_id", "title", "output_type", "is_private", "timestamp"], "type": "object", "properties": {"ai_id": {"title": "Ai Id", "type": "integer"}, "title": {"title": "Title", "type": "string"}, "description": {"title": "Description", "type": "string"}, "output_type": {"title": "Output Type", "type": "string"}, "is_private": {"title": "Is Private", "type": "boolean"}, "python_script_name": {"title": "Python Script Name", "type": "string"}, "python_script_path": {"title": "Python Script Path", "type": "string"}, "timestamp": {"title": "Timestamp", "type": "string", "format": "date-time"}, "last_updated": {"title": "Last Updated", "type": "string", "format": "date-time"}}}, "Body_create_model_file_ai_files_modelfiles_post": {"title": "Body_create_model_file_ai_files_modelfiles_post", "required": ["files"], "type": "object", "properties": {"files": {"title": "Files", "type": "array", "items": {"type": "string", "format": "binary"}}}}, "Body_create_script_file_ai_files_pythonscript_post": {"title": "Body_create_script_file_ai_files_pythonscript_post", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "CreateAI": {"title": "CreateAI", "required": ["title", "output_type", "is_private", "timestamp"], "type": "object", "properties": {"title": {"title": "Title", "type": "string"}, "description": {"title": "Description", "type": "string"}, "output_type": {"title": "Output Type", "type": "string"}, "is_private": {"title": "Is Private", "type": "boolean"}, "timestamp": {"title": "Timestamp", "type": "string", "format": "date-time"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "ShowAI": {"title": "ShowAI", "required": ["ai_id", "title", "output_type", "is_private", "timestamp"], "type": "object", "properties": {"ai_id": {"title": "Ai Id", "type": "integer"}, "title": {"title": "Title", "type": "string"}, "description": {"title": "Description", "type": "string"}, "output_type": {"title": "Output Type", "type": "string"}, "is_private": {"title": "Is Private", "type": "boolean"}, "python_script_name": {"title": "Python Script Name", "type": "string"}, "python_script_path": {"title": "Python Script Path", "type": "string"}, "timestamp": {"title": "Timestamp", "type": "string", "format": "date-time"}, "last_updated": {"title": "Last Updated", "type": "string", "format": "date-time"}}}, "ShowUser": {"title": "ShowUser", "required": ["user_id", "name", "email"], "type": "object", "properties": {"user_id": {"title": "User Id", "type": "string"}, "name": {"title": "Name", "type": "string"}, "email": {"title": "Email", "type": "string"}}}, "UpdateAI": {"title": "UpdateAI", "required": ["title", "output_type", "is_private", "last_updated"], "type": "object", "properties": {"title": {"title": "Title", "type": "string"}, "description": {"title": "Description", "type": "string"}, "output_type": {"title": "Output Type", "type": "string"}, "is_private": {"title": "Is Private", "type": "boolean"}, "last_updated": {"title": "Last Updated", "type": "string", "format": "date-time"}}}, "UpdateUser": {"title": "UpdateUser", "required": ["name", "email"], "type": "object", "properties": {"name": {"title": "Name", "type": "string"}, "email": {"title": "Email", "type": "string"}}}, "User": {"title": "User", "required": ["user_id", "name", "email", "password"], "type": "object", "properties": {"user_id": {"title": "User Id", "type": "string"}, "name": {"title": "Name", "type": "string"}, "email": {"title": "Email", "type": "string"}, "password": {"title": "Password", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}